sortFilter function modal moved to its own module file
